+++ lib/Web2Print/Processor.php
@@ -21,7 +21,7 @@ use Pimcore\Logger;
 use Pimcore\Model;
 use Pimcore\Model\Document;
 use Pimcore\Tool;
-use Pimcore\Web2Print\Processor\PdfReactor8;
+use Pimcore\Web2Print\Processor\PdfReactor;
 use Pimcore\Web2Print\Processor\WkHtmlToPdf;
 use Symfony\Component\Lock\Factory as LockFactory;
 use Symfony\Component\Lock\LockInterface;
@@ -34,7 +34,7 @@ abstract class Processor
     private static $lock = null;
 
     /**
-     * @return PdfReactor8|WkHtmlToPdf
+     * @return PdfReactor|WkHtmlToPdf
      *
      * @throws \Exception
      */
@@ -43,7 +43,7 @@ abstract class Processor
         $config = Config::getWeb2PrintConfig();
 
         if ($config->get('generalTool') === 'pdfreactor') {
-            return new PdfReactor8();
+            return new PdfReactor();
         } elseif ($config->get('generalTool') === 'wkhtmltopdf') {
             return new WkHtmlToPdf();
         } else {
